move all web resources to src folder
